upda

+++ install/tensor-install.py
@@ -2,9 +2,10 @@
 import os
 import sys
 
-if os.path.exists("tensorflow-2.4.4-cp36-cp36m-linux_x86_64.whl") is False:
-   cmd = "wget https://archive.allsky.tv/AMS9/tensorflow-2.4.4-cp36-cp36m-linux_x86_64.whl"
-   os.system(cmd)
+if os.path.exists("tensorflow-installed.txt") :
+   print("python3.6 with tensorflow is already installed.")
+   exit()
+
 if os.path.exists("/usr/bin/python3.6") is False and os.path.exists("/usr/local/bin/python3.6") is False:
    print("NO PYTHON3.6 INSTALL IT.")
    cmd = "sudo add-apt-repository -y ppa:deadsnakes/ppa"
@@ -27,8 +28,11 @@ if os.path.exists("/usr/bin/python3.6") is False and os.path.exists("/usr/local/
    os.system(cmd)
 
 
-cmd = "sudo python3.6 -m pip install /home/ams/amscams/install/tensorflow-2.4.4-cp36-cp36m-linux_x86_64.whl"
-os.system(cmd)
+if os.path.exists("tensorflow-2.4.4-cp36-cp36m-linux_x86_64.whl") is False:
+   cmd = "wget https://archive.allsky.tv/AMS9/tensorflow-2.4.4-cp36-cp36m-linux_x86_64.whl"
+   os.system(cmd)
+   cmd = "sudo python3.6 -m pip install /home/ams/amscams/install/tensorflow-2.4.4-cp36-cp36m-linux_x86_64.whl"
+   os.system(cmd)
 
 cmd = "sudo python3.6 -m pip install tensorboard"
 os.system(cmd)
@@ -56,4 +60,7 @@ os.system(cmd)
 cmd = "sudo python3 -m pip install --upgrade six"
 os.system(cmd)
 cmd = "sudo python3 -m pip install --upgrade requests"
+os.system(cmd)
+
+cmd = "touch tensorflow-installed.txt"
 os.system(cmd)
